requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
gymnasium==0.29.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.6.1
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sb3_contrib==2.6.0
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.5
shapely==2.1.0
six==1.17.0
smmap==5.0.2
stable_baselines3==2.6.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
Werkzeug==3.1.3
yamldataclassconfig==1.5.0
